refactor(rotate_polytopes): extract consts.rs and catalog.rs
